requirements.txt
aiogram>=3.19.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
asyncio>=3.4.3
redis>=5.0.1
pydantic>=2.4.2
python-dotenv>=1.0.0
APScheduler>=3.10.4
pika>=1.3.2
pytest>=7.4.2
pytest-asyncio>=0.21.1
psycopg2-binary>=2.9.9
aiohttp>=3.8.6
loguru>=0.7.2
prometheus-client>=0.17.1
tenacity>=8.2.3
